chore : 테스트 세팅 수정 및 README 개선 (#24)
* [LCR-93] chore: vitest로 테스트 툴 변경

* [LCR-93] chore: Jest 의존성 제거 및 CI 수정

* [LCR-93] chore: 디렉토리 구조 일부 변경

* [LCR-93] chore: README 수정

* [LCR-93] fix: import 경로 수정

* [LCR-93] fix: Home README 수정

* [LCR-93] fix: Test코드가 없다면 skip하도록 스크립트 수정
