+++ package.json
@@ -4,7 +4,8 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "test": "jest",
+    "test": "vitest run --config ./vitest.config.ts --passWithNoTests",
+    "test:storybook": "vitest run --config ./vitest.storybook.config.ts",
     "dev": "vite",
     "build": "tsc -b && vite build",
     "lint": "eslint .",
@@ -19,7 +20,6 @@
     "axios": "^1.8.4",
     "date-fns": "^4.1.0",
     "framer-motion": "^12.8.0",
-    "jest": "^29.7.0",
     "qr-scanner": "^1.4.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
@@ -42,7 +42,6 @@
     "@storybook/test": "^8.6.12",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
-    "@types/jest": "^29.5.14",
     "@types/node": "^22.14.1",
     "@types/react": "^18.3.18",
     "@types/react-dom": "^18.3.5",
@@ -60,15 +59,12 @@
     "eslint-plugin-storybook": "^0.12.0",
     "globals": "^15.14.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "jest-svg-transformer": "^1.0.0",
     "msw": "^2.7.4",
     "playwright": "^1.51.1",
     "prettier": "^3.5.3",
     "storybook": "^8.6.12",
     "storybook-addon-designs": "^6.3.1",
     "tailwindcss": "^4.1.4",
-    "ts-jest": "^29.3.2",
     "ts-node": "^10.9.2",
     "typescript": "~5.6.2",
     "typescript-eslint": "^8.18.2",
