+++ src/apis/README.md
@@ -2,6 +2,16 @@
 
 ---
 
-서버에 필요한 api들을 정의한 디렉토리입니다.
+API 관련 코드를 관리하는 디렉토리입니다.
 
-- 카멜 케이스를 사용하여 정의합니다.
+### 주요 내용
+
+- HTTP 요청 처리를 위한 Axios 인스턴스 설정
+- 백엔드 서버와의 통신을 위한 API 엔드포인트 정의
+
+### Naming Conventions
+
+- 파일명은 파스칼 케이스로 작성합니다.
+- 모든 API 관련 함수와 변수는 카멜 케이스로 작성합니다.
+- API 파일은 기능 또는 도메인별로 구성합니다. (예: AuthApi.ts)
+- HTTP 메서드를 접두어로 사용합니다. (예: getProducts, postOrder)
