+++ src/assets/README.md
@@ -1,8 +1,14 @@
-## 🌠 assets 디렉토리입니다
+## 💡 assets 디렉토리입니다
 
 ---
 
-이미지, 영상을 관리하는 디렉토리입니다.
+이미지, 폰트 등의 파일을 관리하는 디렉토리입니다.
 
-- 이미지 파일은 `logo-manager`와 같이 중간에 `-`를 사용하여 정의합니다.
-- 기본적으로 webp를 사용합니다.
+### 주요 내용
+
+- 이미지는 `.webp` 를 사용하며, 도메인별로 디렉토리를 분리하여 관리합니다.
+
+### Naming Conventions
+
+- 이미지 파일을 제외한 나머지 파일명은 파스칼 케이스로 작성합니다.
+- 이미지 파일은 중간에 `-` 를 사용하여 관리합니다.
