+++ src/components/README.md
@@ -1,8 +1,16 @@
-## 📦 components 디렉토리입니다
+## 💡 components 디렉토리입니다
 
 ---
 
-컴포넌트들을 관리하는 디렉토리입니다.
+재사용이 가능한 컴포넌트들을 관리합니다.
 
-- 디렉토리명은 카멜 케이스를 사용합니다.
-- 파일명은 파스칼 케이스를 사용합니다.
+### 주요 내용
+
+- `@/components/common` : 전역에서 재사용되는 컴포넌트들을 관리합니다.
+- `@/components/boundary` : 바운더리로 사용되는 컴포넌트들을 관리합니다.
+- `@/components/layout` : 레이아웃 컴포넌트들을 관리합니다.
+- `@/components/` : 재사용 가능성이 있는 컴포넌트들을 기능 별로 묶어서 관리합니다.
+
+### Naming Conventions
+
+- 파일명과 함수명은 모두 파스칼 케이스로 작성합니다.
