--- src/fonts/README.md
@@ -1,7 +1,0 @@
-## 💬 fonts 디렉토리입니다
-
----
-
-프로젝트에서 사용하는 폰트를 관리하는 디렉토리입니다.
-
-- 파스칼 케이스를 사용합니다.
