+++ src/index.css
@@ -2,14 +2,14 @@
 
 @font-face {
   font-family: "Pretendard";
-  src: url("./fonts/PretendardVariable.woff2") format("woff2");
+  src: url("./assets/fonts/PretendardVariable.woff2") format("woff2");
   font-weight: 100 900;
   font-display: swap;
 }
 
 @font-face {
   font-family: "GmarketSans";
-  src: url("./fonts/GmarketSansTTFMedium.ttf") format("truetype");
+  src: url("./assets/fonts/GmarketSansTTFMedium.ttf") format("truetype");
   font-display: swap;
 }
