+++ src/router/README.md
@@ -1,29 +1,14 @@
-## ◉ router 디렉토리입니다
+## 💡 router 디렉토리입니다
 
 ---
 
 라우팅 파일을 관리하는 디렉토리 입니다.
 
-### 구현방법
-
-```tsx
-export const Router = createBrowserRouter([
-  {
-    path: "/",
-    element: <GlobalLayout />,
-    children: [
-      {
-        path: "/dashboard",
-        element: <DashBoard />,
-      },
-    ],
-  },
-  {
-    path: "/onboarding",
-    element: <OnBorading />,
-  },
-]);
-```
+### 주요 내용
 
 - createBrowserRouter를 사용하여 라우팅 정보를 관리합니다.
 - 라우팅 경로는 소문자로만 추가합니다.
+
+### Naming Conventions
+
+- 파일과 함수명 모두 파스칼 케이스를 사용합니다.
