+++ src/stores/README.md
@@ -1,0 +1,15 @@
+## 💡 stores 디렉토리입니다
+
+---
+
+Zustand를 사용해서 관리되는 전역 상태 관리 디렉토리입니다.
+
+### 주요 내용
+
+- 클라이언트 상태 관리를 목적으로 합니다.
+- 무분별한 사용은 자제하도록 합니다.
+
+### Naming Conventions
+
+- prefix로 `use`를 사용하고 suffix로 `Store`를 사용합니다.
+- 파일명과 함수명을 동일하게 사용합니다.
