+++ src/types/README.md
@@ -1,0 +1,17 @@
+## 💡 types 디렉토리입니다
+
+---
+
+TypeScript 타입을 정의합니다.
+
+### 주요 내용
+
+- `@/types/api/ApiRequestType.ts` : API 요청에 사용되는 Request DTO Type을 지정합니다.
+  - 모든 Request DTO Type을 해당 파일 내에서 관리하도록 합니다.
+- `@/types/api/ApiResponseType.ts` : API 요청에 사용되는 Response DTO Type을 지정합니다.
+  - 모든 Response DTO Type을 해당 파일 내에서 관리하도록 합니다.
+- `@/types/` : 전역적으로 사용되는 타입들을 정의합니다.
+
+### Naming Conventions
+
+- 파일과 함수명 모두 파스칼 케이스를 사용합니다.
