+++ src/utils/README.md
@@ -1,0 +1,13 @@
+## 💡 types 디렉토리입니다
+
+---
+
+기능 구현에 사용되는 Util 함수들을 관리합니다.
+
+### 주요 내용
+
+- 기능별로 나눠서 관리합니다.
+
+### Naming Conventions
+
+- 파일과 함수명 모두 파스칼 케이스를 사용합니다.
