+++ tsconfig.json
@@ -1,10 +1,10 @@
 {
-  "files": [],
   "references": [
     { "path": "./tsconfig.app.json" },
     { "path": "./tsconfig.node.json" }
   ],
   "compilerOptions": {
+    "jsx": "react-jsx",
     "esModuleInterop": true,
     "baseUrl": ".",
     "paths": {
