+++ vitest.storybook.config.ts
@@ -1,0 +1,27 @@
+// vitest.storybook.config.ts
+import { defineConfig } from "vitest/config";
+import { mergeConfig } from "vite";
+import viteConfig from "./vite.config";
+import path from "node:path";
+import { fileURLToPath } from "node:url";
+import { storybookTest } from "@storybook/experimental-addon-test/vitest-plugin";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+export default mergeConfig(
+  viteConfig,
+  defineConfig({
+    plugins: [storybookTest({ configDir: path.join(__dirname, ".storybook") })],
+    test: {
+      name: "storybook",
+      browser: {
+        enabled: true,
+        headless: true,
+        name: "chromium",
+        provider: "playwright",
+      },
+      setupFiles: [".storybook/vitest.setup.ts"],
+    },
+  }),
+);
